+++ stacks.cpp
@@ -29,7 +29,7 @@ string RuntimeValue::description() const
     case Kind::Null:
         return "null";
     case Kind::FunctionObject:
-        return "functionobj";
+        return "function";
     }
 }
 
@@ -41,6 +41,27 @@ RuntimeValue FunctionObject::call(const vector<RuntimeValue>& arguments) const
     {
         callEnv->variables[parameters[i]] = {arguments[i]};
     }
+    try
+    {
+        body->evaluateNode(callEnv);
+        return {};
+    }
+    catch (const ReturnSignal& r)
+    {
+        return r.value;
+    }
+}
+
+RuntimeValue FunctionObject::call(const vector<RuntimeValue>& arguments, const vector<RuntimeValue>& restArgs) const
+{
+    auto callEnv = std::make_shared<Environment>();
+    callEnv->parent = closure;
+    for (int i = 0; i < parameters.size(); ++i)
+    {
+        callEnv->variables[parameters[i]] = {arguments[i]};
+    }
+    auto arr = std::make_shared<RuntimeValue::Array>(restArgs);
+    callEnv->variables[variadicParamName] = {arr};
     try
     {
         body->evaluateNode(callEnv);